+++ test/entrypoint.sh
@@ -4,7 +4,7 @@ while true
 do
   echo "Waiting for server"
 
-  if curl -isf http://admin:admin@geb_server:15672/api/aliveness-test/%2F > /dev/nul
+  if curl -isf http://admin:admin@rabbitmq:15672/api/aliveness-test/%2F > /dev/nul
   then
     break
   fi